disaster recovery from temp state
This fix helps to recover from a state, where the temporary groups were created, the account groups were not and the original workspac-local groups were deleted.

The fix recreates the original state, with recreating all the workspace-local groups, which does have a temp pair but not an account level pair attached to the workspace.

This helps to recreate the state of the migration with the already existing group_manager.migration_groups_provider