fix/issue_545: Include host in jdbc:/
The UCX tool fetches the 'location' from the table properties while for an external location with jdbc, the host information is stored in 'storage properties'.

Add a field for Table class and fetch Storage Properties from the table. Using the host field from Storage properties, create the full jdbc url