Fix for secret scope apply task always raises ValueError (#683)
### Summary: 
Unfortunately, both the `list_acls` and `get_acl` methods seem to be
very inconsistent for checking whether a permission is applied to a
given secret scope. For this reason, this PR introduces a workaround
where if the permissions don't match, the acl is being applied again.
From my experience of trying multiple workarounds, this was the only one
that proved to be consistent.

**Fixes #627** 

### Changes:
- Added extra safety checks for inflight permission check logic in
secrets.py
- Added integration tests in test_secrets.py

### Tests:
Introduced integration tests for secrets which were missing before.