Added `databricks labs ucx validate-external-locations` command for cli (#715)
Description:
This change contains steps to validate external location defined in the
external tables with the ones already available in the unity external
location. It prints the count of tables that can be migrated for
external location already present. Also shares the list of external
location that needs to be created for the missing ones. It generates a
tf file

Changes:
labs.yml - to add the new command
cli.py - to add code for invoking the new command
locations.py - add a field which gives count of tables using the
external location. also added logic to identify duplicate for jdbc
connection objects, functionality to compare locations and print details
and generate tf file
updated test cases to locations
added new test cases for mapping