Added initial version of `databricks labs ucx migrate-local-code` command (#1067)
The `databricks labs ucx migrate-local-code` command has been added to
facilitate migration of local code to a Databricks environment. This
initial version of the command is highly experimental, with support for
migrating Python and SQL files only. The `.gitignore` file has been
updated to exclude output files and specific configuration files from
being committed to the repository. This command aims to help users and
administrators manage code migration and maintain consistency across
workspaces, while also enhancing the compatibility of local code with
the Unity Catalog, a part of Databricks' offerings for data and AI.