Added external location validation when creating catalogs with `create-catalogs-schemas` command (#1110)
In this release, we have updated the "create\_catalogs\_schemas"
function in "cli.py" to create UC catalogs and schemas based on the
destinations from "create\_table\_mapping". This function previously
used a "CatalogSchema" object created using the workspace client,
installation, and prompts, but has been modified to directly use the
installation object and call the "create\_all\_catalogs\_schemas"
method. We have also added a new feature for catalog validation with
commit message 'Feature/create catalog validation 1101', which includes
the addition of the "create\_all\_catalogs\_schemas" method in
"catalog\_schema.py". The new method generates a list of missing
catalogs and schemas and creates them using the "WorkspaceClient".
Additionally, we have added new imports, modified the constructor of the
"CatalogSchema" class, and added new methods for creating catalogs and
schemas, and updated the unit tests for
"hive\_metastore/catalog\_schema.py". Overall, these changes improve the
robustness and reliability of the tool for working with catalogs and
external locations.

Closes #1101