Improve table migration workflow test and add ACL migration back (#1166)
## Changes
<!-- Summary of your changes that are easy to understand. Add
screenshots when necessary -->
1. Deploy `udfs` table in inventory database when install, which avoids
the potential race condition that two parallel workflow tasks are trying
to create this missing table at the same time.
2. Catch NotFound exception when listing table in schema in
`TablesCrawler._crawl`. During the integration test, schema are being
created and deleted and there is a chance the schema is deleted when
trying list tables inside it. In this case, RuntimeBackend will raise
NotFound and it needs to be handled.
3. Table migration workflow need to use shared mode cluster so it can
crawl table ACLs, in case the `grants` table is empty.
4. Upload wheel to WSFS instead of DBFS if table migration job cluster,
because shared mode cluster cannot use DBFS wheel.
5. Improve table migration workflow integration test:
- updates cluster override to accommodate the new `new_installation `
fixture which by default set cluster override to the default shared mode
cluster.
- use non default name for inventory schema and migrate src schema. So
when the flaky "Schema not found" issue happens again, it would help to
locate the cause of the error.
- catch the NotFound caused by table not migrated and fail the test
instead of retry.
- skip assert on the auto-scale and parallelism config value, until the
prompts are added back. This will be handled in another PR.
- remove `inventory_schema` in `new_installation`, and let the
installation itself create the inventory database. Thus when
`inventory_schema_suffix` is provided, the code will not create two
databases (one with suffix and one whithout suffix). And somehow the
database without suffix is likely to cause flaky "Schema not found"
issue.

### Linked issues

Resolves #1144 
Resolves #1145

---------

Co-authored-by: Serge Smertin <259697+nfx@users.noreply.github.com>