Fixed support for `is_partitioned` property in tables assessment (#1175)
closes #1167

 | Feature | Description |
| --- | --- |
| New Property `is_partitioned` | A new boolean property
`is_partitioned` has been added to the `Table` class in the `tables.py`
module to indicate if a Hive table is partitioned or not. This property
is determined by checking for the presence of the string "# Partition
Information" in the `describe` response from the Hive Metastore. |

| Testing Details | Description |
| --- | --- |
| Integration and Unit Tests | The new functionality includes the
addition of integration and unit tests in the `test_tables.py` file to
ensure the correctness of the `is_partitioned` property. |
| Test Method `test_partitioned_tables` | A new test method,
`test_partitioned_tables`, has been added to check the accuracy of the
`is_partitioned` property for both partitioned and non-partitioned
tables. The method creates four tables, two with partitions and two
without, and then uses the `TablesCrawler` class to retrieve a snapshot
of all the tables. It then asserts that the `is_partitioned` property is
set correctly for each of the tables. |
| Unit and Integration Tests for `is_partitioned` Property and
`TablesCrawler` Object | Appropriate unit and integration tests have
been created to ensure the correct behavior of the new functionality.
This includes creating a mock backend with sample data and checking the
`is_partitioned` property for both partitioned and non-partitioned
tables using the `TablesCrawler` method. |

The new `is_partitioned` property is fully tested and integrated with
existing functionality. The implementation includes thorough testing,
including unit and integration tests, to ensure the correct behavior of
this new functionality.