Added initial version of account-level installer (#1339)
## Changes

Allow admins to run the installer as account admins and install it to
all the workspaces in a given account at once.

- Perform necessary auth to login to the account.
- Prompt for configuration for the first workspace
- Run the installer.
- Confirm that user wants to repeat that for the remaining workspaces
- Repeat
- UCX is installed to all workspaces.

This is done by splitting the prompts for `_configure_new_installation`
into a new method `prompt_for_new_installation`, which saves the answers
to a new `InstallationConfig` data class. This allows the answers to be
reused for other workspaces

### Linked issues
<!-- DOC: Link issue with a keyword: close, closes, closed, fix, fixes,
fixed, resolve, resolves, resolved. See
https://docs.github.com/en/issues/tracking-your-work-with-issues/linking-a-pull-request-to-an-issue#linking-a-pull-request-to-an-issue-using-a-keyword
-->

Resolves #920

### Functionality 

- [x] added relevant user documentation
- [x] modified existing command: `databricks labs install ucx ...` will
switch to account-level installation when the environment variable
`UCX_FORCE_INSTALL` is set to `account`


### Tests
<!-- How is this tested? Please see the checklist below and also
describe any other relevant tests -->

- [x] manually tested