Added Python linter for table creation with implicit format.
Starting with DBR v8.0, the default table format changed from 'parquet' to 'delta'.
This linter yields advice for situations where a 'writeTo', 'table', 'insertInto', or 'saveAsTable' method is invoked, and no '.format(...)' invocation is found in the same chain of calls.
For 'saveAsTable' in particular, the alternative of passing the 'format' as a direct argument is also supported.