Added Python linter for table creation with implicit format (#1435)
## Changes
Starting with DBR v8.0, the default table format changed from 'parquet'
to 'delta'.
This linter yields advice for situations where a 'writeTo', 'table',
'insertInto', or 'saveAsTable' method is invoked, and no '.format(...)'
invocation is found in the same chain of calls.
For 'saveAsTable' in particular, the alternative of passing the 'format'
as a direct argument is also supported.

Only linting is supported for this, no fixing.
If a DRM version is passed to the linter, it will skip linting when this
version is >= 8.0. This is because the format default change from
"parquet" to "delta" starting with v8.0.

Examples:
```
df.writeStream.table("c.db.table") # => linter yields Advice due to implicit format
df.writeStream.format("parquet").table("c.db.table") # => OK!
df.writeStream.saveToTable("c.db.table", format="parquet") # => OK!
```

The linter is implemented in the new file table_creation.py, and makes
use of a few reusable AST utilities in the new file python_ast_util.py.
Unit-tests are added for both source files.

### Linked issues
Relates to #1124 , specifically item 26 -
https://docs.databricks.com/en/archive/runtime-release-notes/8.0.html#delta-is-now-the-default-format-when-a-format-is-not-specified

### Functionality 

- [ ] added relevant user documentation
- [ ] added new CLI command
- [ ] modified existing command: `databricks labs ucx ...`
- [ ] added a new workflow
- [x] modified existing workflow: code migration
- [ ] added a new table
- [ ] modified existing table: `...`

### Tests

- [ ] manually tested
- [x] added unit tests
- [ ] added integration tests
- [ ] verified on staging environment (screenshot attached)

---------

Co-authored-by: Serge Smertin <259697+nfx@users.noreply.github.com>