Fix test failing due to groups not being created yet (#1669)
## Changes
(We think that) sometimes groups are not created before the test run,
e.g.
[this](https://github.com/databrickslabs/ucx/actions/runs/9002466231?pr=1532)
and [that](https://github.com/databrickslabs/ucx/issues/1616). This PR
extends the `make_group` with an optional flag to wait for the groups to
be created before continuing with the test. To limit impact on
performance, we don't wait for the groups to be created by default.

### Linked issues
Hopefully esolves #1616

### Functionality 

- [ ] added relevant user documentation
- [ ] added new CLI command
- [ ] modified existing command: `databricks labs ucx ...`
- [ ] added a new workflow
- [ ] modified existing workflow: `...`
- [ ] added a new table
- [ ] modified existing table: `...`

### Tests
<!-- How is this tested? Please see the checklist below and also
describe any other relevant tests -->

- [x] manually tested
- [ ] added unit tests
- [ ] added integration tests
- [ ] verified on staging environment (screenshot attached)