Remove false warning on DataFrame.insertInto() about the default format changing from parquet to delta (#1823)
## Changes

This pull request removes a spurious incompatible-format warning for
`DataFrameWriter.insertInto()`. The `.insertInto()` operation uses the
existing format of the underlying table; any specified format is
ignored. As such use of this operation be flagged as a potential problem
due to the change of the default format from `parquet` to `delta`.

Incidental changes included in this PR:

 - Removed an unnecessary linting suppression.

### Functionality 

- [ ] added relevant user documentation
- [ ] added new CLI command
- [ ] modified existing command: `databricks labs ucx ...`
- [ ] added a new workflow
- [ ] modified existing workflow: `...`
- [ ] added a new table
- [ ] modified existing table: `...`

### Tests
<!-- How is this tested? Please see the checklist below and also
describe any other relevant tests -->

- [X] manually tested
- [ ] added unit tests
- [ ] added integration tests
- [ ] verified on staging environment (screenshot attached)