Release v0.26.0 (#1865)
* Added migration for Python linters from `ast` (standard library) to
`astroid` package
([#1835](https://github.com/databrickslabs/ucx/issues/1835)). In this
release, the Python linters have been migrated from the `ast` package in
the standard library to the `astroid` package, version 3.2.2 or higher,
with minimal inference implementation. This change includes updates to
the `pyproject.toml` file to include `astroid` as a dependency and bump
the version of `pylint`. No changes have been made to user
documentation, CLI commands, workflows, or tables. Testing has been
conducted through the addition of unit tests. This update aims to
improve the functionality and accuracy of the Python linters.
* Added workflow linter for delta live tables task
([#1825](https://github.com/databrickslabs/ucx/issues/1825)). In this
release, there are updates to the `_register_pipeline_task` method in
the `jobs.py` file. The method now checks for the existence of the
pipeline and its libraries, and registers each notebook or jar library
found in the pipeline as a task. If the library is a Maven or file type,
it will raise a `DependencyProblem` as it is not yet implemented.
Additionally, new functions and tests have been added to improve the
quality and functionality of the project, including a workflow linter
for Delta Live Tables (DLT) tasks and a linter that checks for issues
with specified DLT tasks. A new method,
`test_workflow_linter_dlt_pipeline_task`, has been added to test the
workflow linter for DLT tasks, verifying the correct creation and
functioning of the pipeline task and checking the building of the
dependency graph for the task. These changes enhance the project's
ability to ensure the proper configuration and correctness of DLT tasks
and prevent potential issues.
* Consistent 0-based line tracking for linters
([#1855](https://github.com/databrickslabs/ucx/issues/1855)). 0-based
line tracking has been consistently implemented for linters in various
files and methods throughout the project, addressing issue
[#1855](https://github.com/databrickslabs/ucx/issues/1855). This change
includes removing direct filesystem references in favor of using the
Unity Catalog for table migration and format changes. It also updates
comments and warnings to improve clarity and consistency. In particular,
the spark-table.py file has been updated to ensure that the
spark.log.level is set correctly for UC Shared Clusters, and that the
Spark Driver JVM is no longer accessed directly. The new file,
simple_notebook.py, demonstrates the consistent line tracking for
linters across different cell types, such as Python, Markdown, SQL,
Scala, Shell, Pip, and Python (with magic commands). These changes aim
to improve the accuracy and reliability of linters, making the codebase
more maintainable and adaptable.

Dependency updates:

* Updated sqlglot requirement from <24.2,>=23.9 to >=23.9,<25.1
([#1856](https://github.com/databrickslabs/ucx/pull/1856)).