Release v0.27.1 (#1900)
* Fixed typo in `known.json`
([#1899](https://github.com/databrickslabs/ucx/issues/1899)). A fix has
been implemented to correct a typo in the `known.json` file, an
essential configuration file that specifies dependencies for various
components of the project. The typo was identified in the `gast`
dependency, which was promptly rectified by modifying an incorrect
character. This adjustment guarantees precise specification of
dependencies, thereby ensuring the correct functioning of affected
components and maintaining the overall reliability of the open-source
library.