Group migration: continue permission migration even if one or more groups fails (#1924)
## Changes

This PR updates permission migration so that if an error occurs while
migrating the permissions for a group it moves onto the next group
instead of stopping immediately. Any errors are raised at the end
instead.

At present the information about which groups succeeded and failed is
logged but not persisted in any other way.

### Linked issues

Progresses #1914.

### Functionality 

- [X] modified existing workflow: `group-migration`

### Tests

- [X] added unit tests