Group renaming: wait for consistency before completing task (#1944)
## Changes

Renaming groups is eventually consistent. This PR updates the task that
renames workspace groups so that it does not complete until the renaming
is visible via direct get on the group and via group enumeration.

The integration tests already dealt with the eventual consistency
problem, but downstream tasks in the workflow might also run into
problems and this wasn't being handled. Ensuring that renaming has
completed before the task prevents this from being an issue.

### Functionality 

- [X] modified existing workflow: `group-migration`

### Tests

- [X] added unit tests

---------

Co-authored-by: Cor <jczuurmond@protonmail.com>