Revise integration test timeoutsIntegration test timeouts (#1945)
## Changes

Some of our integration tests have been set up to retry when specific
failures occur make them more resilient in the face of timeouts and race
conditions. One of the retry settings is a timeout, which does not
interrupt a running test but rather specifies the deadline after which a
new retry attempt will not be made.

For some of these tests, a single attempt can take longer than the
configured timeout, especially when running over VPN. This defeats the
purpose of the retry setup for that test: it's there because failures
are known to occur, but because the attempt takes longer than the
timeout deadline a second attempt will never be made.

This PR updates the timeouts for tests where the single attempt was
noted as being within 20% (or longer) of the deadline. Increasing the
timeouts means that the retry behaviour will be more likely to have its
intended effect.

---------

Co-authored-by: Serge Smertin <259697+nfx@users.noreply.github.com>