Update purge suffix so that resources are preserved for at least as long as the configured timeout. (#2265)
## Changes

This PR updates the purge suffix used by integration tests so that
resources are marked for preservation _at least_ as long as the
configured timeout. The situation is that:

- Many resources can be marked with a tag (or a special suffix in the
name) to ensure they are preserved until a specific time.
- The timestamp refers to a UTC-based hour, such as: `2024072915`. This
means that resources can be deleted at any moment after 3 p.m. (UTC).
- Prior to this change, resources created as late as 2:44:59 p.m. would
be tagged for deletion with `2024072915` meaning deletion as of 3 p.m.
This is 0:15 later instead of the intended 1:15.

The default timeout for tests was 1:15 in an attempt to deal with
resources being deleted when the test started near the hour boundary.
This PR also updates the default timeout to 1:00 because the extra 0:15
isn't needed to ensure this.