Integration tests: purge suffixes and `RemoveAfter` tags for fixtures that don't already have this (#2388)
## Changes

This PR updates all remaining integration test fixtures so that where
possible any resources they create have a `RemoveAfter` tag or the
equivalent suffix on their name. (This prevents the resource from being
removed from the test environment while the test is underway.)

Doing this in a single PR is intended to reduce future effort tracking
down spurious test failures when this occurs.

The main remaining resource that is often deleted while integration
tests are underway is the folder for the UCX installation itself.
Unfortunately the name of this folder is fixed and tightly coupled to
the generated product-name being used for the test. Dealing with this is
non-trivial and has been omitted from this PR.

### Linked issues

Supersedes #2369.

### Tests

- integration tests