Fix failing integration test: `test_reflect_account_groups_on_workspace_skips_groups_that_already_exists_in_the_workspace` (#2624)
## Changes
Also check for **workspace** groups to exists with the same name. Follow
up on #2615

In hindsight, what might have happened before: the workspace group was
not created before the account group was reflected in the workspace,
thus the test passes. With the changes in PR #2615, we got more
consistent behaviour because we wait for the workspace group to exists
before running the to-be-tested code. However, the code did not handle
that situation. This PR handles when a workspace group has the same name
as an to-be-reflected-in-the-workspace account group

### Linked issues

Resolves #2623

### Functionality

- [x] modified existing workflow: `group-migration`

### Tests

- [x] modified integration tests:
`test_reflect_account_groups_on_workspace_skips_account_groups_when_a_workspace_group_has_same_name`