Update the workspace content cache used during linting (#2953)
## Changes

This PR began as an effort to retrofit the `WorkspaceCache` (and
associated implementation) with type hints to help ensure correctness.
This proved to be quite difficult due to the dynamic nature of the cache
and binary vs. text modes. As such, this PR updates the `WorkspaceCache`
in the following ways:

- We now only cache the binary content of workspace files. This
simplifies the implementation substantially and avoids correctness
issues with respect to various options that can be specified to control
decoding as binary content. (Decoding is well optimised compared to
upstream/downstream processing, so this is a reasonable tradeoff.)
- The cache now resolves paths, ensuring that different ways of
referring to the same content are handled as a single cache entry.
- We now raise an error immediately if a non-absolute path is requested:
workspace paths are always absolute, so this will catch errors earlier.
- For the paths where we check for a BOM-marker prior to reading as
text, if the opened stream for the path is seekable then we only open it
once instead of twice.

### Tests

- added (and updated existing) unit tests
- existing integration tests